fix(tests): cast memory requested/released counters to u128 in e2e checks
Lease memory request/release counters widened to u128; remaining e2e
assertions still compared them to u64 setup/reservation values and
broke compile-compat.

- fs-diffreal-e2e AsBuiltInvocationPlan / leaf_evidence_matches: cast
  setup.memory and requested_memory when checking memory_requested_bytes
  and memory_released_bytes on children and receipts
- fs-flagship-e2e production_scale sparse LBM: cast retained_lease_bytes
  for requested_bytes on live and after-drop lease receipts

Tests only; peak/used/limit comparisons already matched the wider type.

+++ crates/fs-diffreal-e2e/src/invocation.rs
@@ -214,13 +214,13 @@ impl AsBuiltInvocationPlan {
             && children[0].direct_consumed() == setup.consumed
             && children[0].direct_memory_peak_bytes() == setup.memory
             && children[0].memory_peak_bytes() == transaction_memory_peak
-            && children[0].memory_requested_bytes() == setup.memory
-            && children[0].memory_released_bytes() == setup.memory
+            && children[0].memory_requested_bytes() == setup.memory as u128
+            && children[0].memory_released_bytes() == setup.memory as u128
             && children[0].output_retained_bytes() == 0
             && receipt.remaining() == expected_remaining
             && receipt.memory_peak_bytes() == transaction_memory_peak
-            && receipt.memory_requested_bytes() == requested_memory
-            && receipt.memory_released_bytes() == requested_memory
+            && receipt.memory_requested_bytes() == requested_memory as u128
+            && receipt.memory_released_bytes() == requested_memory as u128
             && receipt.output_retained_bytes() == retained_output
     }
 }
@@ -250,8 +250,8 @@ fn leaf_evidence_matches(child: &fs_exec::ChildReceipt, expected: ExpectedLeafEv
         && child.direct_consumed() == expected.consumed
         && child.direct_memory_peak_bytes() == expected.memory
         && child.memory_peak_bytes() == expected.memory
-        && child.memory_requested_bytes() == expected.memory
-        && child.memory_released_bytes() == expected.memory
+        && child.memory_requested_bytes() == expected.memory as u128
+        && child.memory_released_bytes() == expected.memory as u128
         && child.output_retained_bytes() == expected.consumed.output().get()
 }
